Move RPC initialization to its own class
